test rest controleur article
création rest controlleur Article -- Test Postman OK

Modification des entités pour inclure les relation entre commande bon article.

+++ src/main/java/com/gestionets/Gestion/Entreprise/Service/impl/ArticleServiceImpl.java
@@ -32,12 +32,12 @@ public class ArticleServiceImpl implements ArticleService {
     }
 
     @Override
-    public void SupprimerArticle(Integer id) {
+    public void SupprimerArticle(Long id) {
         articleRepository.deleteById(id);
     }
 
     @Override
-    public Optional<Article> RechercheArticleparID(Integer id) {
+    public Optional<Article> RechercheArticleparID(Long id) {
         return articleRepository.findById(id);
     }
 }
